Some classes moved into more suitable naming packages, refactored paths in classes where they are mentioned
